통계정보_계산서발행금액

+++ sberp/src/main/java/egovframework/sts/dto/StsDTO.java
@@ -14,7 +14,7 @@ public class StsDTO {
   private String stsDt;
   
   @ApiModelProperty(value="계산서발행 금액 ", example = "")
-  private int billAmt;
+  private long billAmt;
   @ApiModelProperty(value="납부완료 금액 ", example = "")
   private long payAmt;
